+++ src/imars3d/backend/corrections/ring_removal.py
@@ -4,7 +4,7 @@
 import logging
 import param
 import multiprocessing
-from imars3d.backend.util.util import clamp_max_workers
+from imars3d.backend.util.functions import clamp_max_workers
 import scipy
 import numpy as np
 import tomopy.util.mproc as mproc
