+++ src/imars3d/backend/diagnostics/rotation.py
@@ -5,7 +5,7 @@ import logging
 import numpy as np
 
 import param
-from imars3d.backend.util.util import clamp_max_workers
+from imars3d.backend.util.functions import clamp_max_workers
 import multiprocessing
 from concurrent.futures import ProcessPoolExecutor
 from multiprocessing.managers import SharedMemoryManager
