+++ src/imars3d/backend/diagnostics/tilt.py
@@ -4,7 +4,7 @@
 import logging
 import param
 import multiprocessing
-from imars3d.backend.util.util import clamp_max_workers
+from imars3d.backend.util.functions import clamp_max_workers
 import numpy as np
 from typing import Tuple
 from functools import partial
