+++ src/imars3d/backend/io/data.py
@@ -4,7 +4,7 @@ import re
 import logging
 import param
 import multiprocessing
-from imars3d.backend.util.util import clamp_max_workers
+from imars3d.backend.util.functions import clamp_max_workers
 import numpy as np
 import tifffile
 from functools import partial
