+++ src/imars3d/backend/preparation/normalization.py
@@ -3,7 +3,7 @@
 """iMars3D normalization module."""
 import logging
 import param
-from imars3d.backend.util.util import clamp_max_workers
+from imars3d.backend.util.functions import clamp_max_workers
 import multiprocessing
 import numpy as np
 import tomopy
